// src/app/opmobilitybackend/models/get-warehouse-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { TownResponse } from '../models/town-response';
export interface GetWarehouseResponse {

  /**
   * Address of the warehouse
   */
  address: string;

  /**
   * ID of the warehouse
   */
  id: number;

  /**
   * Name of the warehouse
   */
  name: string;

  /**
   * Town of the warehouse
   */
  town: TownResponse;
}